+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "forms-typed",
+  "name": "forms-typed-examples",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
