+++ projects/forms/CHANGELOG.md
@@ -1,0 +1,6 @@
+# 1.0.0 
+Initial release
+# 1.0.0-rc.4
+Add documentation README.md and CHANGELOG.md
+# 1.0.0-rc.3
+Include Typed Forms, Control Value Accessor Connector and forEachControlIn
