+++ projects/forms/tsconfig.lib.prod.json
@@ -1,5 +1,8 @@
 {
   "extends": "./tsconfig.lib.json",
+  "compilerOptions": {
+    "importHelpers": false
+  },
   "angularCompilerOptions": {
     "enableIvy": false
   }
